Move basic_agent to the old directory
